+++ test/ui/pages.js
@@ -29,6 +29,7 @@ const pages = [
   ['help', () => require('../../out/ui/webview/helpHtml.js').getHelpHtml(webview, { fsPath: '/ext' })],
   ['stats', () => require('../../out/ui/webview/statsHtml.js').getStatsHtml(webview)],
   ['taskBoard', () => require('../../out/ui/webview/taskBoardHtml.js').getTaskBoardHtml(webview)],
+  ['calendar', () => require('../../out/ui/webview/calendarHtml.js').getCalendarHtml(webview)],
   ['relatedNotesDebug', () => require('../../out/ui/webview/relatedNotesDebugHtml.js')
       .getRelatedNotesDebugHtml(webview, {
         filePath: 'notes/a.md', sourceLine: 1, title: 'Entry', tags: [],